+++ addOns/ascanrulesBeta/src/main/java/org/zaproxy/zap/extension/ascanrulesBeta/XxeScanRule.java
@@ -167,16 +167,6 @@ public class XxeScanRule extends AbstractAppPlugin implements ChallengeCallbackP
         return Alert.RISK_HIGH;
     }
 
-    @Override
-    public void init() {
-        if (ChallengeCallbackImplementor.getExtensionCallback() == null) {
-            // The callback extension is not available, cant do anything :(
-            getParent()
-                    .pluginSkipped(
-                            this, Constant.messages.getString(MESSAGE_PREFIX + "nocallback"));
-        }
-    }
-
     /**
      * Scan rule to check for XXE vulnerabilities. It checks both for local and remote using the ZAP
      * API and also a new model based on parameter substitution
@@ -198,29 +188,37 @@ public class XxeScanRule extends AbstractAppPlugin implements ChallengeCallbackP
             // using an external bouncing site, in this case we use
             // the ZAP API as a server for the vulnerability check
             // using a challenge/response model based on a random string
-            //
-            String challenge = randomString(CHALLENGE_LENGTH);
 
-            try {
-                // Prepare the attack message
-                msg = getNewMsg();
-                payload = getCallbackAttackPayload(challenge);
-                msg.setRequestBody(payload);
+            // Skip XXE Remote File Inclusion Attack when the callback extension is not available.
+            if (ChallengeCallbackImplementor.getExtensionCallback() != null) {
+                String challenge = randomString(CHALLENGE_LENGTH);
 
-                // Register the callback for future actions
-                callbackImplementor.registerCallback(challenge, this, msg);
+                try {
+                    // Prepare the attack message
+                    msg = getNewMsg();
+                    payload = getCallbackAttackPayload(challenge);
+                    msg.setRequestBody(payload);
 
-                // All we need has been done...
-                sendAndReceive(msg);
+                    // Register the callback for future actions
+                    callbackImplementor.registerCallback(challenge, this, msg);
 
-            } catch (IOException ex) {
-                // Do not try to internationalise this.. we need an error message in any event..
-                // if it's in English, it's still better than not having it at all.
-                log.warn(
-                        "XXE Injection vulnerability check failed for payload ["
-                                + payload
-                                + "] due to an I/O error",
-                        ex);
+                    // All we need has been done...
+                    sendAndReceive(msg);
+
+                } catch (IOException ex) {
+                    // Do not try to internationalise this.. we need an error message in any event..
+                    // if it's in English, it's still better than not having it at all.
+                    log.warn(
+                            "XXE Injection vulnerability check failed for payload ["
+                                    + payload
+                                    + "] due to an I/O error",
+                            ex);
+                }
+
+                // Exit if the scan has been stopped
+                if (isStop()) {
+                    return;
+                }
             }
 
             // Check if we've to do only basic analysis (only remote should be done)...
@@ -229,14 +227,6 @@ public class XxeScanRule extends AbstractAppPlugin implements ChallengeCallbackP
             }
 
             // Check #2 : XXE Local File Reflection Attack
-            // ------------------------------------------------------
-            // This attack is not described anywhere but the idea is
-            // very simple: use the original XML request and substitute
-            // every content and attribute with a fake entity which
-            // include a sensitive local file. If the page goes in error
-            // or reflect and manage the sent content you can probably
-            // have the file included in the HTML page and you can check it
-            //
             localFileReflectionAttack(getNewMsg());
 
             // Check if we've to do only medium sized analysis (only remote and reflected will be
@@ -245,42 +235,36 @@ public class XxeScanRule extends AbstractAppPlugin implements ChallengeCallbackP
                 return;
             }
 
+            // Exit if the scan has been stopped
+            if (isStop()) {
+                return;
+            }
+
             // Check #3 : XXE Local File Inclusion Attack
-            // ------------------------------------------------------
-            // This attack is described in
-            // https://owasp.org/www-community/vulnerabilities/XML_External_Entity_(XXE)_Processing
-            // trying to include a local file and maybe have the inclusion back in
-            // the result page. This situation is very uncommon because it works
-            // only in case of a bare XML parser which execute the conetnt and then
-            // gives it back almost untouched (maybe because it applies an XSLT or
-            // query it using XPath and give back the result).
             localFileInclusionAttack(getNewMsg());
         }
     }
 
-    // TODO: Add Doc Comments
-    protected void localFileReflectionAttack(HttpMessage msg) {
-        Matcher matcher;
-        String localFile;
-        String response;
+    /**
+     * Local File Reflection Attack substitutes every attribute in the original XML request with a
+     * fake entity which includes a sensitive local file. The attack is repeated for every file
+     * listed in the LOCAL_FILE_TARGETS. The response returned is pattern matched against
+     * LOCAL_FILE_PATTERNS. An alert is raised when a match is found.
+     *
+     * @param msg new HttpMessage with the same request as the base. This is used to build the
+     *     attack payload.
+     */
+    private void localFileReflectionAttack(HttpMessage msg) {
         String payload = null;
         try {
             String requestBody = createLfrPayload(msg.getRequestBody().toString());
-
             for (int idx = 0; idx < LOCAL_FILE_TARGETS.length; idx++) {
-                // Prepare the message
-                localFile = LOCAL_FILE_TARGETS[idx];
+                String localFile = LOCAL_FILE_TARGETS[idx];
                 payload = MessageFormat.format(requestBody, localFile);
-                // msg = getNewMsg();
                 msg.setRequestBody(payload);
-
-                // Send message with local file inclusion
                 sendAndReceive(msg);
-
-                // Parse the result
-
-                response = msg.getResponseBody().toString();
-                matcher = LOCAL_FILE_PATTERNS[idx].matcher(response);
+                String response = msg.getResponseBody().toString();
+                Matcher matcher = LOCAL_FILE_PATTERNS[idx].matcher(response);
                 if (matcher.find()) {
 
                     newAlert()
@@ -290,16 +274,10 @@ public class XxeScanRule extends AbstractAppPlugin implements ChallengeCallbackP
                             .setMessage(msg)
                             .raise();
                 }
-
-                // Check if the scan has been stopped
-                // if yes dispose resources and exit
                 if (isStop()) {
-                    // Dispose all resources
-                    // Exit the rule
                     return;
                 }
             }
-
         } catch (IOException ex) {
             log.warn(
                     "XXE Injection vulnerability check failed for payload ["
@@ -309,27 +287,30 @@ public class XxeScanRule extends AbstractAppPlugin implements ChallengeCallbackP
         }
     }
 
-    // TODO: Add Doc Comments
-    protected void localFileInclusionAttack(HttpMessage msg) {
-        String localFile;
-        String response;
-        Matcher matcher;
+    /**
+     * Local File Inclusion Attack is described in
+     * https://owasp.org/www-community/vulnerabilities/XML_External_Entity_(XXE)_Processing. The
+     * attack builds a payload for every file listed in LOCAL_FILE_TARGETS with the ATTACK_HEADER
+     * and the ATTACK_BODY. The response returned is pattern matched against LOCAL_FILE_PATTERNS. An
+     * alert is raised when a match is found.
+     *
+     * <p>This situation is very uncommon because it works only in case of a bare XML parser which
+     * execute the content and then returns the content almost untouched (maybe because it applies
+     * an XSLT or query it using XPath and give back the result)
+     *
+     * @param msg new HttpMessage with the same request as the base. This is used to build the
+     *     attack payload.
+     */
+    private void localFileInclusionAttack(HttpMessage msg) {
         String payload = null;
         try {
             for (int idx = 0; idx < LOCAL_FILE_TARGETS.length; idx++) {
-                // Prepare the message
-                localFile = LOCAL_FILE_TARGETS[idx];
+                String localFile = LOCAL_FILE_TARGETS[idx];
                 payload = MessageFormat.format(ATTACK_HEADER + ATTACK_BODY, localFile);
-                // msg = getNewMsg();
                 msg.setRequestBody(payload);
-
-                // Send message with local file inclusion
                 sendAndReceive(msg);
-
-                // Parse the result
-
-                response = msg.getResponseBody().toString();
-                matcher = LOCAL_FILE_PATTERNS[idx].matcher(response);
+                String response = msg.getResponseBody().toString();
+                Matcher matcher = LOCAL_FILE_PATTERNS[idx].matcher(response);
                 if (matcher.find()) {
 
                     newAlert()
@@ -339,16 +320,10 @@ public class XxeScanRule extends AbstractAppPlugin implements ChallengeCallbackP
                             .setMessage(msg)
                             .raise();
                 }
-
-                // Check if the scan has been stopped
-                // if yes dispose resources and exit
                 if (isStop()) {
-                    // Dispose all resources
-                    // Exit the rule
                     return;
                 }
             }
-
         } catch (IOException ex) {
             // Do not try to internationalise this.. we need an error message in any event..
             // if it's in English, it's still better than not having it at all.